update PATCH, DELETE

+++ src/modules/user/user.service.ts
@@ -1,4 +1,11 @@
-import { Inject, Injectable, LoggerService, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import {
+    ForbiddenException,
+    Inject,
+    Injectable,
+    LoggerService,
+    NotFoundException,
+    UnauthorizedException
+} from '@nestjs/common';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { DatabaseService } from '../database/database.service';
 import { I18nContext, I18nService } from 'nestjs-i18n';
@@ -95,12 +102,16 @@ export class UserService {
         }
     }
 
-    async updateUserById(id: string, dto: UpdateUserDto): Promise<User> {
+    async updateUserById(id: string, dto: UpdateUserDto, user: User): Promise<User> {
         try {
-            const user = await this.getUserById(id);
+            const userById = await this.getUserById(id);
+
+            if (userById.id !== user.id) {
+                throw new ForbiddenException('Нельзя обновить чужой аккаунт');
+            }
 
             return this.prisma.user.update({
-                where: { id: user.id },
+                where: { id: userById.id },
                 data: dto
             });
         } catch (error) {
@@ -109,9 +120,13 @@ export class UserService {
         }
     }
 
-    async deleteUserById(id: string): Promise<User> {
+    async deleteUserById(id: string, user: User): Promise<User> {
         try {
-            const user = await this.getUserById(id);
+            const userById = await this.getUserById(id);
+
+            if (userById.id !== user.id) {
+                throw new ForbiddenException('Нельзя обновить чужой аккаунт');
+            }
 
             return this.prisma.user.delete({
                 where: { id: user.id }
